shared_mutex: return typed exceptional future in with_* error handlers
Currently, exceptions in templates for functions that may
throw on move are handled in a generic way using
current_exception_as_future().

But if the passed functor returns a value, we need
to use current_exception_as_future<value_type> instead
so the error-free and error case returned future type
will match.

This change fixes that and adds unit test reproducers.

Fixes #1295

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>