http: Add url encoding method
As per RFC the path?a=b&c=d strings should be properly encoded not to
include spaces and other "bad" characters in it. From the API
perspective it's inconvenient to ask user for properly formatter query
string. This patch adds the respective methods and a test, that query
formatting can be parsed by (existing) parser producing the original
result.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>