file_desc: add spawn_actions_add_{close, dup2}() helpers
to pave the road to add a wrapper for make_pipe(), which is in turn
will be used to refactor spawn_process(), so it can be simplified
with the make_pipe() helper.

this is not a must have since we have the accessor exposing
the fd. but this change follows the convention of file_desc which
encapsulate the underlying POSIX calls.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>