sockets: Wait on socket peer closing the connection
This patch adds the connected_socket::wait_input_shutdown() call that
returns a future that resolves when the connection is terminated by the
peer or when the local socket is shutdown for reading. This event can
be waited on even if a client doesn't read from (or write to) its end
while server SUDDENLY closes it.

This can be useful for keep-alive HTTP sockets. When a client session is
created and a request is served, client may want to cache the connected
socket without sending new requrests right at once, but still needs to
know if the server recycles the connection on his side.

The returned future may resolve with data left in the socket or in the
input stream obtained with sk.input(), the caller is supposed to handle
it on its own. Can be called at most once per socket.

Implementation depends on the socket type (and reactor backend)

- Posix socket

There's a useful polling event in the Linux called POLLRDHUP which is
reported when socket is shutdown(SHUT_RD) or peer of the connection
closes or shutdown(SHUT_RW). Two last options cause FIN to be sent to
the client, respectively kernel sets it upon FIN receiving.

Said that, implementation is in extending the pollable_fd_state with the
poll_rdhup() method that asks the backend to poll for it. AIO and URing
backends just submit the respective iocb into the kernel, epoll backend
needs a bit more care because it mixes events in one mask and needs to
"parse" it when event comes.

- Native stack

The implementation mimics Linux kernel -- the connection promise is
resolved when it's being read-aborted or when the connection receives
FIN packet.

- TLS

This just forwards the call to the underlying socket.

Test included

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>