docker: bump up to clang {14,15} and gcc {11,12}
since we only support the latest two major releases of compilers.
at the moment of writing, Clang just released v15, and the latest
major release of GCC is v12.

so we should install them respectively.

because ubuntu jammy does not ship clang-15. we need to bump up
the base image from ubuntu:jammy to ubuntu:kinetic despite that
kinetic is not an LTS release. we should use ubuntu 24.04 once
it's out.

[avi: squash related circleci change, regenerate toolchain,
      and use versioned image tag instead of :latest]