build: check for fix of CWG2631
CWG 2631 (https://cplusplus.github.io/CWG/issues/2631.html) reports
an issue on how the default argument is evaluated. this problem is
more obvious when it comes to how `std::source_location::current()`
is evaluated as a default argument. but not all compilers have the
same behavior, see https://godbolt.org/z/PK865KdG4.

notebaly, clang-15 evaluates the default argument at the callee
site. so we need to check the capability of compiler and fall back
to the one defined by util/source_location-compat.hh if the compiler
suffers from CWG 2631.

in this change, FindSourceLocation.cmake is reused for checking
the behavior of compiler by adding a new test which fails if the
expected line number is different from the one returned by
`std::source_location::current()`.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>