allocator: add an option to reserve additional memory for the OS
This patch adds an option to specify the memory reserved for the
OS not as a flat value, but as a modifier for the recommended
amounts.

The intended use case of this option is when the application
is using both seastar and system allocators and it relies on
the seastar defaults for the reservation of memory reserved
for the OS but wants to make sure that there is enough memory
not managed by seastar allocator for both the application
and the rest of the system.

The optimal value for this option depends on the implementation
of the specific application, so this option is marked "internal"
and can be only set programmaticaly using the app_template config.

If the user specifies the reserve-memory while this option is set
by the application, the memory reserved for the system will be a
sum of these values. As a result, there will be enough memory for
system and non-seastar application allocations, and the seastar
allocator will take control of the remaining memory.

If the user specifies the memory option while the additional
memory option is set by the application, the memory managed
by the seastar allocator will be the difference of these
values, and there will be an extra "additional memory" reserved
for the system.

Signed-off-by: Wojciech Mitros <wojciech.mitros@scylladb.com>