build: define GnuTLS_{LIBRARIES, INCLUDE_DIRS} only if GnuTLS is found
this change has not actual impact on the behavior of the building
process. it's just more correct. as per the convention, a
Find<Package>.cmake is not supposed to define <Package>_LIBRARIES
and <Package>_INCLUDE_DIRS unless the package is found.

actually, we could just drop GnuTLS_{LIBRARIES, INCLUDE_DIRS}, as
we link against GnuTLS using the imported target, but let's just
keep them at this moment in case projects using Seastar take advantage
of these defined variables and actually are using them.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>