tcp: Close connection when getting RST from server
While trying to connect to another server address, if the server response an [RST,ACK] segment, the connect() function still being hang and failed just on timed out. due to no connection were closed and _connection_done has not be setted.
According to RFC 9293, Receiving RST:
If the receiver was in any other state except SYN-RECEIVED or LISTEN, it aborts the connection and advises the user and goes to the CLOSED state.

Closes #1339

Fixes #1334.

[avi: adjust commit message from pull request cover letter, add Fixes
      annotation]