build: use URING_LIBRARIES for liburing_libs in .pc file
the generator expression of $<TARGET_LINKER_FILE:tgt> only works if
the `tgt` in question is an executable or a library. if the `tgt`
is an interface library, we would have to use something like

$<TARGET_PROPERTY:URING::uring,INTERFACE_LINK_LIBRARIES>

but there are two cases here:

1. URING::uring is imported as an INTERFACE library
2. URING::uring is imported as an IMPORTED library

it'd be complicated to check if the target's TYPE using a generator
expression. so let's just use `URING_LIBRARIES` for the value of
`liburing_libs` in the .pc file.

this change paves the road to the refactory of FindLibUring.cmake
to use pkg_search_module(... IMPORTED_TARGET GLOBAL ...), which in
turn simplifies the FindLibUring.cmake to import the library found
by target_link_libraries().

Signed-off-by: Kefu Chai <tchaikov@gmail.com>