reactor: define __NR_pidfd_open if not defined
there is chance that we need to compile Seastar with an outdated
glibc or kernel-headers package, and hence don't have access the
macro definition of SYS_pidfd_open. in either case, we need to
define __NR_pidfd_open so Seastar can compile. even if the kernel
does not support this syscall, it's not the end of world -- we'll
fall back to a loop of waitpid().

this change should address the FTBFS on, for instance, an up to date
CentOS8 installation with stock glibc.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>

Closes #1342