http: Convert http::request::content from sstring to temporary_buffer<char>
S3 upload doesn't support chunked-encoded bodies, instead it requires
"classical" enccoding with content-length header. However, the
http::request is not ready for this. In order to write body the caller
has two options:

- provide body-writer callback. In this case the request will be sent
  with chunked encoding
- provide ready to use sstring. In this case client will send the
  content-length header and the raw body

None is suitable for s3 client. Using body writer callback just won't
work, using sstring body is quite ugnly, because strictly speaking it's
not going to be string at all.

Good news is that string body is deprecated for server-side and was
recently re-used by experimental client, so it should not be dangerous
to convert content from sstring to temporary buffer.

For convenience, writing body as sstring is still supported, just the
string is converted to buffer.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>