Add an API for the metrics layer to manipulate metrics dynamically.
It is typical for software that builds on top of Seastar report
per-shard metrics.  With modern hardware, this can quickly grow to a
large number of metrics that overload the reporting system and the
collecting server.

One of the main issues around metrics reduction is that many of the
metrics are only helpful in certain situations.

It will be best to have the option to add metrics freely and choose in
runtime what to report.

Changing label values lets a metrics-collecting server (e.g.,
Prometheus) ask for only some of the metrics.

Disabling and enabling the metrics will ensure they will not be
calculated and reported at all, but the configuration will need to be
changed when required.

This series adds an option to perform metrics manipulation dynamically.
It includes adding, removing, and changing labels and also enable,
disable metrics, and enable and disable the skip_when_empty option.

The general logic follows Prometheus metrics_relabel_config
configuration.

The relabel config is a vector of rules applied one after the other.
Rules can be changed at any point, it would first update all the
existing metrics, and then any new metrics will be updated according to
the rules.

Here are a few examples:

To add a level label with a value 1, to the reactor_utilization metric:
    std::vector<sm::relabel_config> rl(1);
    rl[0].source_labels = {"__name__"};
    rl[0].target_label = "level";
    rl[0].replacement = "1";
    rl[0].expr = "reactor_utilization";
   set_relabel_configs(rl);

To report only the metrics with the level label equals 1

    std::vector<sm::relabel_config> rl(2);
    rl[0].source_labels = {"__name__"};
    rl[0].action = sm::relabel_config::relabel_action::drop;

    rl[1].source_labels = {"level"};
    rl[1].expr = "1";
    rl[1].action = sm::relabel_config::relabel_action::keep;
    set_relabel_configs(rl);