build: export the full path of uring with URING_LIBRARIES
to address the same issue fixed by
3db15b5681c460a759641a3907eb9db84bdcf395. otherwise we'd pass "uring"
to the compiler when linking against liburing, and it would refuse to
work. as "uring" alone is not a valid command line option to the
compiler, "-luring" is.

this change addresses the regression introduced by
35b2dc3145b636f9476fa3d5e8be9b998a0f4623, and partially reverts it.
but this change only uses LibUring_LIBDIR as the HINTS when finding the
library, as LibUring_LIBDIR is more specific and it should be identical
to LibUring_LIBRARY_DIRS, as only a single moduleSpec is passed to
pkg_check_modules().

Signed-off-by: Kefu Chai <tchaikov@gmail.com>