http: Add content_length_data_sink implementation
This sink is mirrors the content_length_source_impl thing. It accepts
input buffers and forwards them to the output taking care than not more
than the given amount of bytes flow into the target.

After the stream is closed, the creator's variable is updated with the
actual number of bytes written by the stream user. Next patch will make
use of it to detect body "underflow".

If the callback tries to generate more bytes than it was suggested with
the content-length value, the sink::put method resolves in exceptional
future.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>