http: Add request::write_body() functor-based method
There are two options how to generate request body:
- provide the body sstring. In this case the string will be sent into
  the body in plain encoding and the content-length header will be
  automatically added to request headers
- provide the body-generating function. In this case the function will
  be called with the chunk-encoding sink, so that the body is
  automatically split into chunks and chunks are correctly encoded into
  the output stream

This patch adds the third option -- to generate the body with a
function, but when its size is for some reason known in advance. In this
case, the body will be sent plain-encoded and content-length will be
set. The content_length_data_sink is used, so if the function generates
more bytes than it had to, the resulting request will still be correct.

If the provided function didn't generate enough bytes and closed the
stream earlier, the whole request sending is aborted with exceptional
future.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>