http: Add connection::skip_body() sugar
Sometimes the request maker isn't interested in reply body, but only
needs status and/or headers from it. Example -- S3 multipart upload
completion request. The reply would contain parts summary which the
caller may not need to check. However, leaving the body unread is not
good if the connection is keep-alive and re-used for next request. In
this case the left body will be read back as if it was the next reply
itself which's wrong.

The added function drains the input stream clearing it for the next
request.

Non-keep-alive requests can just close the socket.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>