build: use different toolset base on specified compiler
before this change, we just use a guessed toolset when building
Boost if we build Seastar with
`configure.py --cook Boost --compiler clang++-16`.
there is good chance that the guessed toolset is not the one
specified in the command line. please search "$TOOLSET" in
https://github.com/boostorg/boost/blob/master/bootstrap.sh
for more details.

after this change, two popular compilers are mapped to its
corresponding toolset, so we are closer to cooking Boost
with specified compiler. the future improvement would be
to customize the .jam file to customize the path of compiler
and compiling options with the user specified ones.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>