build: detect debug build of yaml-cpp also
if yaml-cpp is built with -DCMAKE_BUILD_TYPE=Debug, the generated
library would be named like "libyaml-cppd.so".

before this change, we only try to find yaml-cpp when searching for
its library. but if this library is built with -DCMAKE_BUILD_TYPE=Debug,
it won't be found. and hence the built might fail if the system does not
have yaml-cpp installed.

after this change, both the release and debug builds are detected
and exposed. select_library_configurations() is used to choose and
the library variables according to which one(s) is found.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>