http: Add client class
The class is going to hold a bunch of cached keep-alive connections to
server and call make_request()s on them. Since the connection becomes
"request-scoped", the way to handle response changed.

The connection::make_request() returned future<reply> and the caller was
supposed to get connection.in(resp) to read the reply body. With the
client::make_request() it's not going to work the same way, as
client.in() will be hard to map to specific underlying connection. So
instead the client::make_request() accepts a noncopyable function
handler that will be called with reply and input_stream with body.

To facilitate requests handling the client::make_request also accepts
optional 'expected' reply status code.

For now connections are not keep-alived, this patch just introduces the
client::make_request() wrapper.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>