http: Cache connections on client
This patch adds the boost intrusive double-linked list of connections on
client. When the client::make_request() asks for a connection it first
tries to withdraw one from that list. If not found, the new connection
is made using connection factory.

When the connection is put back it's cached in list and (!) the
shared_ptr is held by the connection on itself. This makes sure the
connection is not released (when it is it's auto-unlinked from the list,
but it doesn't happen in this patch and will come later).

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>