http: Track remote connection termination
When the connection is cached on the client it can get closed by the
server. In this case the client should know this early and release the
connection's resources.

To make it work the connected_socket.wait_input_shutdown() future is
waited upon. The continuation holds shared pointer on the connection, so
when it resolves the connection is released and is auto-unlinked from
the pool.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>