tests: only destroy allocated pending connections
this change should address the regression introduced by
6cd5f9fb1cbbe90acf9f63305c4c412e765e34fa.

before this change, in
`loopback_connection_factory::destroy_all_shards()`, we use the current
shard id as the index in `_pending`, and submit `destroy_shard()` to all
shards, but there are chances that the Seastar application is configured
with more shards than loopback_connection_factory is. in
6cd5f9fb1cbbe90acf9f63305c4c412e765e34fa, we limit the loopback connection
factory sharding to 1, but in general, we run the tests with two shards,
hence we have the heap-buffer-overflow failure reported by ASan after
that commit.

after this change,

* `destroy_all_shards()` only destroys the pending connections on allocated shards.
* and assert() is added to ensure we don't have out-of-bound access with
  Debug build.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>