Adding relabel config tests
This patch adds tests for the relabel_config functionality.

It consists of three tests:
test_relabel_add_labels
Test that you can add a new label based on name.

test_relabel_drop_label_prevent_runtime_conflicts
Test that you can remove a label. It also validates that if a metric
conflict occurs (i.e., after removing a label, there is a conflict) a
label is added to resolve the conflict.  It also validates that cleaning
the configuration return things to normal.

test_relabel_enable_disable_skip_when_empty
Test that the skip_when_empty can be set and unset.  It also validates
the option to disable all metrics besides a few based on labels (i.e.,
the use case that only metrics with a specific label value will be
reported).

Signed-off-by: Amnon Heiman <amnon@scylladb.com>