build: enable non-library targets using default option value
before this change, the non-library targets are always disabled even
if Seastar is embedded using `add_subdirectory()` and these targets
are enabled with commands like:

set(Seastar_TESTING ON CACHE BOOL "" FORCE)

in the parent project. the value of `Seastar_TESTING` is overridden
by Seastar.

but there are chances that the parent project still want to enable
one or more of these targets. in this case, it'd be desirable to
keep the value(s) set by the parent project.

after this change, instead of overriding the cached values, we
provide the default value for them depending on if Seastar is
a top-level project. So the parent project can now set the value
of options like `Seastar_APPS`. and these values are preseved.

Fixes #1380
Signed-off-by: Kefu Chai <tchaikov@gmail.com>