reactor: s/resize(0)/clear()/
from functionality perspective, resize(0) is identical to clear().
but static_vector::resize(0) could use more CPU cycles than the latter.
for instance, it compares the new size with the old size to determine
if it should value-initialize the new elements if any.
while clear() just nukes all the elements and set the new size to 0.
it should be faster than resize(0). and more importantly, it has
better readability.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>