http, httpd: optimize header field assignment
In the request parser, we lookup the field name twice, even in the
simple case: once to check if it already exists, and the second time
to insert it.

Optimize by using try_emplace to merge the check and insertion into
a single lookup.

Apply the same logic to the response parser too.

Closes #1395