sstring: specialize uninitialized_string() using 'if constexpr'
before this change, we define a default implmentation of
uninitialized_string for std::basic_string, and define a partial
specialization for basic_sstring, but partial specialization cannot be
picked by the compiler if it compiles something like

auto s = uninitialized_string();

instead, the compiler always go with the first template. but
it is less efficient, as it is desgined for std::basic_string, and
might create a temporary sstring and throw it away by assigning it
with a new instance if the specified size is greater than the size
of the default-constructed one. this is not ideal and beat the
purpose of uninitialized_string(). it is supposed to be efficient.

after this change, a helper template class is defined to
tell of a given class is a specialization of basic_sstring, so we
can use it to specialize `uninitialized_string()` in its implementation
without relying on a partial specialization for this template function.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>