reactor: allow registering handler multiple times for a signal.
std::unordered_map::empalce() inserts a new element into the
container constructed in-place with the given args only if there
is no element with the key in the container, which means we can
only register handler for a specific signal only once and that's
not reasonable.

We often use the stop_signal(which should have been a standard
component of seastar but not yet) to manage SIGINT/SIGTERM for
gracefully stopping application. It waits until receiving SIGINT
or SIGTERM and then register a no-op handler on destruction to
unregister the handler registered on construction, which however
cannot work properly since the handler can be registered only once
for a signal before this patch.

Signed-off-by: Jianyong Chen <baluschch@gmail.com>