reactor: use a for loop for submitting ios
before this change, a while loop is used for submitting i/o requests
in aio_storage_context. it's not very obvious how the iterator is
incremented.

after this change, a for loop is used. and iocbs is used as an iterator
to advance in _submission_queue, so the
aio_storage_context::submit_work() can have a better readability.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>