Added safe pointer utility
A safe pointer is a mechanism for interacting with a possibly destructed object:
Sometimes some object A needs to interact with another object B, e.g., A calls one of B's methods.
If B happens to be already destructed (and freed), this results in a use-after-free bug.
One way to avoid this problem is for A to manage the lifetime of B via a unique_ptr or shared_ptr.
Sometimes, however, it is undesirable to have such ownership relation, and instead A would like to
just skip the interaction in case B has already died (in the above example, A would simply not call
the method on B).

This class provides a solution;
in the above exaple, B is given a safe_ptr_factory as class member:

class B {
  public:
    safe_ptr_factory<B> sp_factory;
    B() : sp_factory(this) {}
};

safe_ptr<B> sp;
{
  B b;
  sp = b.sp_factory.get_safe_ptr();
} // b destructs here

*sp; // throws exception