Fix spurious wakeups in perf-based stall detector
The perf based stall detector has a spurious wakeup problem.

The underlying cause is that re-arming the timer used an ioctl
PERF_EVENT_IOC_RESET, which one might assume resets the event
count associated with overflow and the subsequent signal.
However, that is not the case: it resets the event *count* but not
the separately-tracked overflow counter.

See  _perf_event_reset in kernel/events/core.c and note that
event->count is not used to trigger overflows.

Due to this, the "simple" path for arm_timer() was effectively a no-op
and the timer would fire on the existing periodic schedule. This
causes frequent spurious wakeups when the timer is frequently
re-armed in start_task_run: when tasks are not stalling, we should
prevent signals entirely since we keep bumping the timer period
up, but in fact we get signals every period as above.

This change switches to use PERF_EVENT_IOC_PERIOD which changes
the period but also forces a reset of the current overflow period to
start from now.

After this change, I do not see any spurious signals, but I do keep
detection for spurious signals "just in case" and also because the reset
behavior of this ioctl appears on all archs only in 3.14 (and on Arm
in 3.7). We use reactor::now() to detect spurious signals rather than
reading from the perf_event _fd since this avoids the need to reset
the event count on every arming.

This change should improve performance (especially tail latency)
primarily because we avoid frequent spurious signals.