Reset stall suppression state on reporter change
We have unit tests that check that the right number of reactor stalls
are reported after running some workloads which should stall. These
rely on replacing the reporting function to track non-suppressed stalls.

However, we don't reset the suppression state between tests, so any
test after the first will report zero stalls. This happens to work
for the current set of tests because only one test produces non-zero
expected stalls, so only that test triggers suppression, but the check
that the other tests incur zero stalls is spurious: it may succeed
simply because suppression is in effect.

Change the reporter method to also reset the suppression state so they
start from a clean(ish) slate.