Add user and kernel spin tests
Add some tests to the stall detector test which spin in userspace and
the kernel respectively. These have no asserts but are meant for
manual verification of the backtraces: that there is 0 and 1 kernel
backtrace for the user and kernel tests respectively.

We do not assert because the test is not exact: there is a small chance
that the spin loop happens to be in the "other" ring when the backtrace
is taken.

Issue #1412.
Issue #1413.