In perf-based stall detector, clear ring buffer
The perf-based stall detector sets up a ring-buffer which collects
kernel samples (if the thread is in the kernel at the moment of
interrupt), which may be reported by the perf-based stall detector when
a stall occurs.

However, we didn't read from the queue except when a stall was detected
so in general the ring buffer would fill up, which caused two problems:

 - When a stall was reported a large number of associated kernel stacks
   might be reported at the same time, but logically the current
   stack (at the moment of interruption) should have 0 or 1 kernel
   stacks.
 - The ring buffer may fill up long before a stall occurs, so the
   reported stacks may be from much earlier in time and unrelated to
   the current stall.

This change fixes by this cleaning the ring buffer every time we arm
the timer, so we expect to find exactly 1 stack (which may be empty
if the interrupt happened in userspace) when the stall detector fires
again.

Fixes #1412.