file: Generalize in/out streams making
In order to make input/output stream for a file there are two helpers --
make_file_(input|output)_stream out there. Both make file-based sink or
source, but these sink and source implementations silently assume that
the file is a file-on-disk. Strictly speaking, these are compatible with
posix_file_impl, but not with generic file. E.g. pipes (if they exist)
might break on read-ahead and write-behind.

This patch adds a customization point to file_impl so that custom file
implementations could generate specific sink/source for the stream.

One of the next users of it would be S3 file, which supports posix-ish
data_source, but will require very specific data_sink with append-only
and extra buffering of 5Mb scattered chunks.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>