Dump memory diagnostics at error level on abort
When seastar is configured to throw on allocation failure, we have a
bunch of rules which determine whether a memory diagnosis report
is output, to avoid spamming the log in the case of repeated failures.
In general this means that most allocation failures are logged at debug
log level, with some "critical" failures being logged at error.

However, when --abort-on-seastar-bad-alloc is set, there will be at most
one allocation failure in the lifetime of the process, after which will
we abort so we'd *really* like the diagnosis to appear before we
abort: essentially all allocation failures are "critical".

Enforce this by always logging at error log level when a failure occurs
that will cause an abort.