Allow explicit dumping of memory diagnostics
We trigger a dump of memory diagnostics to the log in certain cases,
such as when a critical allocation fails, or when a fatal OOM occurs.

Add an internal method which allows triggering this manually, used only
purposes: allowing us to call this method from a test so that we can
check that it does not allocate.