Test that memory diagnostics do not allocate
This adds a change that asserts that dumping memory diagnostics don't
allocate. Allocating during diagnostics dumping is dangerous since
these are often dumped in low-or-no memory situations, so additional
allocations during the dump which fail may at best terminate the
dump and at worse obscure the true error.

This test uses the memory statistics to assert that no allocations
occur during 1000 invocations of the method to log the memory
diagnostics.

Currently this test fails occasionally since we allocation when the
second on the logging clock changes.