Avoid allocating during logging
Ideally, the logging path doesn't log, because logging is important
during situations like low-memory, where diagnostics may be output. If
logging tries to allocate at that point, we may get a nested OOM with
less-than-desirable results.

A test which checks if logging will allocate in the context of memory
logging diagnostics shows that it does allocate: once per second when
the second hand of the clock "ticks over" as we apply an optimization
to avoid formatting the time more than once per second, which allocates
a std::string.

Change this to use log_buf with a static buffer of 32 bytes which should
be "enough for everyone" (and if it is not, it simply allocates) to make
up the difference.