reactor: disable io_uring on older kernels if not enough lockable memory is available
Before Linux 5.12[1], io_uring accounted the memory of the ring buffers
as locked memory. This amounts to 32k/shard in our configuration, and
as the default mlock limit was 64k at the time, it allows for a maximum
of two vcpus.

To prevent problems, disable io_uring for those old kernels if less
than 8MB of lockable memory is available. This 8MB is more than
enough, but also allows other applications (e.g. mail clients that
lock encryption keys) to co-exist with Seastar. In practice lockable
memory will be either 64k (old configurations), 8MB (newer kernels
and/or systemd) or infinite (server installations that choose to let
the Seastar application lock its memory), so we aren't missing anything
by disallowing 7.3MB.

[1] https://github.com/torvalds/linux/commit/26bfa89e25f42d2b26fe951bbcf04bb13937fbba