seastar/coroutine: drop std-coroutine.hh
now that we only support the latest two major versions of GCC and
Clang, which are GCC {11,12} and Clang {14,15} respectively. all of
them support C++20 coroutines instead of coroutines-ts, which was
considered experimental before C++20 is officialy announced. and
before C++20, the coroutines header file was located at
`experimental/coroutine`, also we had to pass `-fcoroutines` or
`-fcoroutines-ts` to compiler to enable this feature explicitly.

but since C++20 coroutines is a part of C++20 standard, `-std=c++20`
enables C++20 coroutines. we don't need to detect it by checking
the compiler version or to check the existence of
`experimental/coroutine` anymore, neither do we need to patch
the standard library for a working coroutines support.

so, in this change

* we don't detect the existence of `experimental/coroutine`
* we don't patch `std::experimental` to get the coroutines work

C++17 + coroutines support is intact, as both GCC-11 and Clang-14
support C++20 coroutines if `-fcoroutines` or `-fcoroutines-ts` is
passed to in command line options when compiling in C++17 mode.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>