reactor: move convert_hex_safe() into print_safe.hh
back in 008289160272e339318e8049a19bf1c57ea08154, `convert_hex_safe()`
was added in place of `convert_zero_padded_hex_safe()` to save the space
in the printed backtrace. but this new function is almost identical
to the old one apart from the padding character.

in this change, `convert_hex_safe()` is moved to `util/print_safe.hh`,
so the functions with similar function are grouped together.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>