scheduling: add a `rename` callback to scheduling_group_key_config
Per-scheduling-group values might want to know the name of their
scheduling group, e.g. if they hold some metrics for the group.
But scheduling groups can be renamed, and currently there is no
way to propagate that info to the values.

This patch allows the user of per-scheduling-group values to pass
a callback called when the scheduling group is renamed.
As with its sibling callbacks (`constructor` and `destructor`) it
is called in the relevant scheduling group, so the value can learn
the new name of its group using `current_scheduling_group()`.