test: Add RPC timeout test cases
Check two options -- that messages get properly time-out-ed on a live
RPC connection, and that messages get properly time-out-ed on an RPC
connection that's in the process of connecting/negotiating.

Be extremely cautious and check not only that the sending resloves with
timeout exception, but also that it resolves _before_ actual blocking
disappears.

The latter case fails currently, it will be fixed with the next patch.

refs: #1446

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>