rpc: Keep dummy frame in the outgoing queue until negotiated
Front entry in the queue is uncancellable. That's not perfect, but is
something we have to live with. However, there's one corner case when
this can be fixed -- during the connect+negotiate phase cancelling or
timeouting the packet can work. In fact it should work, because it used
to work before c437920c ("Replace RPC outgoing queue with continuation
chain", so it's actually the regression.

To make it happen client queues a dummy message in front of the outgoing
queue queue, so that any other subsequent sending would happen to be at
least the 2nd one which can be withdrawn from the queue. Once negotiated
the dummy entry goes away and does't interfere with the packets flow,
neither the front entry stays cancellable.

fixes: #1446

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>