reactor: set_shard_field_width() after resource::allocate()
in debug build, if `--smp 0` is passed to seastar, we'd have following
error:

```
seastar/src/core/reactor.cc:4139:35: runtime error: -inf is outside the range of representable values of type 'unsigned int'
```

but in the very same function where the runtime error is thrown,
`resource::allocate(rc)` is called. this function also checks for 0
number of processors, and throw better runtime_error of
"number of processing units must be positive".

in this change, we set the logger's shard field width after `smp::count`
is checked for zero. so we can have a better error message.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>